updated example wrapper and SUT
