+++ resources/harness_adaptive/hierons_fsm.fsm
@@ -1,0 +1,10 @@
+0 0 0 1
+0 0 1 3
+0 1 1 3
+1 0 0 1
+1 1 1 3
+2 0 1 3
+2 1 0 0
+2 1 1 2
+3 0 0 2
+3 1 0 0
