+++ resources/harness_adaptive/hierons_fsm.in
@@ -1,0 +1,2 @@
+a
+b
