+++ resources/harness_adaptive/hierons_fsm.out
@@ -1,0 +1,2 @@
+0
+1
