+++ resources/harness_adaptive/hierons_fsm.state
@@ -1,0 +1,4 @@
+S1
+S2
+S3
+S4
